+++ src/pull-requests/pull-request.service.ts
@@ -173,8 +173,11 @@ export class PullRequestService {
       .distinct()
       .select("pull_requests.author_login", "author_login")
       .addSelect("MAX(pull_requests.updated_at)", "updated_at")
+      .addSelect("users.id", "user_id")
       .innerJoin("repos", "repos", `"pull_requests"."repo_id"="repos"."id"`)
-      .addGroupBy("author_login");
+      .innerJoin("users", "users", `"pull_requests"."author_login"="users"."login"`)
+      .addGroupBy("author_login")
+      .addGroupBy("users.id");
 
     const filters = this.filterService.getRepoFilters(pageOptionsDto, startDate, range);