+++ src/user/user.module.ts
@@ -10,6 +10,9 @@ import { RepoFilterService } from "../common/filters/repo-filter.service";
 import { EndorsementService } from "../endorsement/endorsement.service";
 import { DbEndorsement } from "../endorsement/entities/endorsement.entity";
 import { ApiServicesModule } from "../common/services/api-services.module";
+import { DbUserList } from "../user-lists/entities/user-list.entity";
+import { DbInsight } from "../insight/entities/insight.entity";
+import { TierModule } from "../tier/tier.module";
 import { DbUser } from "./user.entity";
 import { UserService } from "./services/user.service";
 import { UserController } from "./user.controller";
@@ -32,6 +35,7 @@ import { UserEndorsementController } from "./user-endorsement.controller";
 import { UserRecommendationController } from "./user-recommendation.controller";
 import { DbUserOrganization } from "./entities/user-organization.entity";
 import { UserFollowingController } from "./user-following.controller";
+import { UserOrganizationService } from "./user-organization.service";
 
 @Module({
   imports: [
@@ -47,12 +51,15 @@ import { UserFollowingController } from "./user-following.controller";
         DbRepo,
         DbEndorsement,
         DbUserOrganization,
+        DbInsight,
+        DbUserList,
       ],
       "ApiConnection"
     ),
     PullRequestModule,
     RepoModule,
     ApiServicesModule,
+    TierModule,
   ],
   controllers: [
     UserController,
@@ -75,7 +82,15 @@ import { UserFollowingController } from "./user-following.controller";
     UserNotificationService,
     UserCollaborationService,
     EndorsementService,
+    UserOrganizationService,
   ],
-  exports: [UserService, UserHighlightsService, UserFollowService, RepoService, EndorsementService],
+  exports: [
+    UserService,
+    UserHighlightsService,
+    UserFollowService,
+    RepoService,
+    EndorsementService,
+    UserOrganizationService,
+  ],
 })
 export class UserModule {}